# docs/swagger.yaml
basePath: /Vehicle
definitions:
...
  swagger.TechnicalServiceType:
    properties:
      name:
        type: string
    type: object
  swagger.Technician:
    properties:
      dni:
        type: integer
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
host: 192.168.40.104:60030
info:
  contact: {}
  version: "1"
paths:
...
  /technicalServiceType/:
    get:
      description: Obtiene una lista de todos los Tipos de Servicios Tecnicos.
      parameters:
      - default: 1
        description: Número de página para la paginación
        in: query
        name: page
        type: integer
      - default: 10
        description: Tamaño de página para la paginación
        in: query
        name: pageSize
        type: integer
      - description: Nombre del Tipo de servicio Tecnico
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene todos los Tipos de Servicios Tecnicos
      tags:
      - TechnicalServiceType
    post:
      consumes:
      - application/json
      description: Crea un nuevio tipo de servicio tecnico con los datos proporcionados
      parameters:
      - description: Datos de la marca de vehículo a crear
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.TechnicalServiceType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea una nuevo tipo de servicio tecnico con los datos proporcionados
      tags:
      - TechnicalServiceType
  /technicalServiceType/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un Tipo de Servicio Tecnico  de la base de datos basado
        en el ID proporcionado
      parameters:
      - description: ID del Tipo de Servicio tecnico a eliminar
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mensaje de éxito al eliminar
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tipo de  Servicio Tecnico no encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Elimina un Tipo de Servicio Tecnico por ID
      tags:
      - TechnicalServiceType
    get:
      description: Obtiene un Tipo de Servicio Tecnico por su ID específico
      parameters:
      - description: ID del Tipo de servicio Tecnico
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalles del Tipo de servicio técnico
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: ID inválido'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Tipo de servicio tecnico no encontrado'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene Tipo de Servicio Tecnico específico
      tags:
      - TechnicalServiceType
    put:
      consumes:
      - application/json
      description: Actualiza un Tipo de Servicio tecnico por ID con los datos proporcionados
      parameters:
      - description: ID del tipo de servicio Tecnico a actualizar
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del Tipo de Servicio Tecnico a actualizar
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.TechnicalServiceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Actualiza un Tipo de Servicio Tecnico por ID
      tags:
      - TechnicalServiceType
  /technician/:
    get:
      description: Obtiene una lista de todos los Tecnicos.
      parameters:
      - default: 1
        description: Número de página para la paginación
        in: query
        name: page
        type: integer
      - default: 10
        description: Tamaño de página para la paginación
        in: query
        name: pageSize
        type: integer
      - description: Nombre del Tecnico
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene todos los Tecnicos
      tags:
      - Technicians
    post:
      consumes:
      - application/json
      description: Crea un nuevio técnico con los datos proporcionados
      parameters:
      - description: Datos de la marca de vehículo a crear
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.Technician'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swagger.Technician'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea una nuevo técnico con los datos proporcionados
      tags:
      - Technicians
  /technician/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un Tecnico  de la base de datos basado en el ID proporcionado
      parameters:
      - description: ID del tecnico a eliminar
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mensaje de éxito al eliminar
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Dispositivo Avl no encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Elimina un Tecnico por ID
      tags:
      - Technicians
    get:
      description: Obtiene un Tecnico por su ID específico
      parameters:
      - description: ID del Tecnico
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalles del técnico
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: ID inválido'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: tecnico no encontrado'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene Tecnico específico
      tags:
      - Technicians
    put:
      consumes:
      - application/json
      description: Actualiza un tecnico por ID con los datos proporcionados
      parameters:
      - description: ID del Tecnico a actualizar
        in: path
        name: id
        required: true
        type: integer
      - description: Datos de la marca de vehículo a actualizar
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.Technician'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Actualiza un Tecnico por ID
      tags:
      - Technicians
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"